Refactor: Move global/missing functions to own file
