+++ functions/global_functions.php
@@ -1,0 +1,79 @@
+<?php
+
+/**
+ * detect missing gettext and fake function
+ */
+if(!function_exists('gettext')) {
+	function gettext ($text) 	{ return $text; }
+	function _($text) 			{ return $text; }
+}
+
+/**
+ * create links function
+ *
+ *	if rewrite is enabled in settings use rewrite, otherwise ugly links
+ *
+ *	levels: $el
+ */
+function create_link ($l0 = null, $l1 = null, $l2 = null, $l3 = null, $l4 = null, $l5 = null, $l6 = null ) {
+	# get settings
+	global $User;
+
+	// url encode all
+	if(!is_null($l6))	{ $l6 = urlencode($l6); }
+	if(!is_null($l5))	{ $l5 = urlencode($l5); }
+	if(!is_null($l4))	{ $l4 = urlencode($l4); }
+	if(!is_null($l3))	{ $l3 = urlencode($l3); }
+	if(!is_null($l2))	{ $l2 = urlencode($l2); }
+	if(!is_null($l1))	{ $l1 = urlencode($l1); }
+	if(!is_null($l0))	{ $l0 = urlencode($l0); }
+
+
+	# set normal link array
+	$el = array("page", "section", "subnetId", "sPage", "ipaddrid", "tab");
+	// override for search
+	if ($l0=="tools" && $l1=="search")
+    $el = array("page", "section", "ip", "addresses", "subnets", "vlans", "ip");
+
+	# set rewrite
+	if($User->settings->prettyLinks=="Yes") {
+		if(!is_null($l6))		{ $link = "$l0/$l1/$l2/$l3/$l4/$l5/$l6"; }
+		elseif(!is_null($l5))	{ $link = "$l0/$l1/$l2/$l3/$l4/$l5/"; }
+		elseif(!is_null($l4))	{ $link = "$l0/$l1/$l2/$l3/$l4/"; }
+		elseif(!is_null($l3))	{ $link = "$l0/$l1/$l2/$l3/"; }
+		elseif(!is_null($l2))	{ $link = "$l0/$l1/$l2/"; }
+		elseif(!is_null($l1))	{ $link = "$l0/$l1/"; }
+		elseif(!is_null($l0))	{ $link = "$l0/"; }
+		else					{ $link = ""; }
+
+		# IP search fix
+		if ($l0=="tools" && $l1=="search" && isset($l2) && substr($link,-1)=="/") {
+    		$link = substr($link, 0, -1);
+		}
+	}
+	# normal
+	else {
+		if(!is_null($l6))		{ $link = "index.php?$el[0]=$l0&$el[1]=$l1&$el[2]=$l2&$el[3]=$l3&$el[4]=$l4&$el[5]=$l5&$el[6]=$l6"; }
+		elseif(!is_null($l5))	{ $link = "index.php?$el[0]=$l0&$el[1]=$l1&$el[2]=$l2&$el[3]=$l3&$el[4]=$l4&$el[5]=$l5"; }
+		elseif(!is_null($l4))	{ $link = "index.php?$el[0]=$l0&$el[1]=$l1&$el[2]=$l2&$el[3]=$l3&$el[4]=$l4"; }
+		elseif(!is_null($l3))	{ $link = "index.php?$el[0]=$l0&$el[1]=$l1&$el[2]=$l2&$el[3]=$l3"; }
+		elseif(!is_null($l2))	{ $link = "index.php?$el[0]=$l0&$el[1]=$l1&$el[2]=$l2"; }
+		elseif(!is_null($l1))	{ $link = "index.php?$el[0]=$l0&$el[1]=$l1"; }
+		elseif(!is_null($l0))	{ $link = "index.php?$el[0]=$l0"; }
+		else					{ $link = ""; }
+	}
+	# prepend base
+	$link = BASE.$link;
+
+	# result
+	return $link;
+}
+
+/**
+ * Escape HTML and quotes in user provided input
+ * @param  mixed $data
+ * @return string
+ */
+function escape_input($data) {
+       return empty($data) ? '' : htmlentities($data, ENT_QUOTES);
+}
